Refactoring the server/service packages
Advantages:

- `Interface` becomes `Service` (better naming)
- `Service` (formerly `Interface`) is declared where it is actually used
- No more generic `service` package name
- No more repetition (`service.NewDomainService` becomes `domain.NewService`)
